refactor transformers interviews into a modal wrapper over a local module
